requirements.txt
Flask>=2.0
flask-restx>=0.5.1
Flask-JWT-Extended>=4.3.1
Flask-SQLAlchemy>=2.5
python-dotenv>=0.21.0
Faker>=18.0.0
Flask-Cors>=3.0